improvement: mail structure
no issue
- in preparation for subscribers V2
- do not implement code in index.js
- create mail utils
