+++ core/server/api/authentication.js
@@ -3,7 +3,8 @@ var _                = require('lodash'),
     pipeline         = require('../utils/pipeline'),
     dataProvider     = require('../models'),
     settings         = require('./settings'),
-    mail             = require('./mail'),
+    mail             = require('./../mail'),
+    apiMail          = require('./mail'),
     globalUtils      = require('../utils'),
     utils            = require('./utils'),
     errors           = require('../errors'),
@@ -171,7 +172,7 @@ authentication = {
                     '/ghost/reset/' +
                     globalUtils.encodeBase64URLsafe(data.resetToken) + '/';
 
-            return mail.generateContent({
+            return mail.utils.generateContent({
                 data: {
                     resetUrl: resetUrl
                 },
@@ -189,7 +190,7 @@ authentication = {
                     }]
                 };
 
-                return mail.send(payload, {context: {internal: true}});
+                return apiMail.send(payload, {context: {internal: true}});
             });
         }
 
@@ -413,7 +414,7 @@ authentication = {
                 ownerEmail: setupUser.email
             };
 
-            return mail.generateContent({data: data, template: 'welcome'})
+            return mail.utils.generateContent({data: data, template: 'welcome'})
                 .then(function then(content) {
                     var message = {
                             to: setupUser.email,
@@ -428,7 +429,7 @@ authentication = {
                             }]
                         };
 
-                    mail.send(payload, {context: {internal: true}}).catch(function (error) {
+                    apiMail.send(payload, {context: {internal: true}}).catch(function (error) {
                         errors.logError(
                             error.message,
                             i18n.t(
