+++ core/server/api/mail.js
@@ -1,35 +1,26 @@
 // # Mail API
 // API for sending Mail
-var _             = require('lodash').runInContext(),
-    Promise       = require('bluebird'),
+
+var Promise       = require('bluebird'),
     pipeline      = require('../utils/pipeline'),
-    config        = require('../config'),
     errors        = require('../errors'),
-    GhostMail     = require('../mail'),
+    mail          = require('../mail'),
     Models        = require('../models'),
     utils         = require('./utils'),
     notifications = require('./notifications'),
-    path          = require('path'),
-    fs            = require('fs'),
-    templatesDir  = path.resolve(__dirname, '..', 'mail', 'templates'),
-    htmlToText    = require('html-to-text'),
-    readFile      = Promise.promisify(fs.readFile),
     docName       = 'mail',
     i18n          = require('../i18n'),
     mode          = process.env.NODE_ENV,
     testing       = mode !== 'production' && mode !== 'development',
     mailer,
-    mail;
-
-_.templateSettings.interpolate = /{{([\s\S]+?)}}/g;
+    apiMail;
 
 /**
  * Send mail helper
  */
-
 function sendMail(object) {
-    if (!(mailer instanceof GhostMail) || testing) {
-        mailer = new GhostMail();
+    if (!(mailer instanceof mail.GhostMailer) || testing) {
+        mailer = new mail.GhostMailer();
     }
 
     return mailer.send(object.mail[0].message).catch(function (err) {
@@ -58,7 +49,7 @@ function sendMail(object) {
  * @typedef Mail
  * @param mail
  */
-mail = {
+apiMail = {
     /**
      * ### Send
      * Send an email
@@ -95,9 +86,9 @@ mail = {
         }
 
         tasks = [
-                utils.handlePermissions(docName, 'send'),
-                send,
-                formatResponse
+            utils.handlePermissions(docName, 'send'),
+            send,
+            formatResponse
         ];
 
         return pipeline(tasks, options || {});
@@ -127,7 +118,7 @@ mail = {
          */
 
         function generateContent(result) {
-            return mail.generateContent({template: 'test'}).then(function (content) {
+            return mail.utils.generateContent({template: 'test'}).then(function (content) {
                 var payload = {
                     mail: [{
                         message: {
@@ -158,45 +149,7 @@ mail = {
         ];
 
         return pipeline(tasks);
-    },
-
-    /**
-     *
-     * @param {Object} options {
-     *              data: JSON object representing the data that will go into the email
-     *              template: which email template to load (files are stored in /core/server/mail/templates/)
-     *          }
-     * @returns {*}
-     */
-    generateContent: function (options) {
-        var defaults,
-            data;
-
-        defaults = {
-            siteUrl: config.forceAdminSSL ? (config.urlSSL || config.url) : config.url
-        };
-
-        data = _.defaults(defaults, options.data);
-
-        // read the proper email body template
-        return readFile(path.join(templatesDir, options.template + '.html'), 'utf8').then(function (content) {
-            var compiled,
-                htmlContent,
-                textContent;
-
-            // insert user-specific data into the email
-            compiled = _.template(content);
-            htmlContent = compiled(data);
-
-            // generate a plain-text version of the same email
-            textContent = htmlToText.fromString(htmlContent);
-
-            return {
-                html: htmlContent,
-                text: textContent
-            };
-        });
     }
 };
 
-module.exports = mail;
+module.exports = apiMail;
